# Projekti/Projekat1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Data Service",
        "description": "Express.js service as Data Service"
    },
    "host": "localhost:3333",
    "basePath": "/",
    "tags": [
        {
            "name": "Vitals"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/getVitals": {
            "get": {
                "tags": [
                    "Vitals"
                ],
                "summary": "Get vital parameters for specific user",
                "parameters": [
                    {
                        "name": "userID",
                        "type": "integer",
                        "in": "query",
                        "description": "User ID",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "User not found."
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/postVitals": {
            "post": {
                "tags": [
                    "Vitals"
                ],
                "summary": "Post vital parameters for specific user.",
                "description": "Sample request:\r\n    \r\n    POST /Gateway\r\n    {\r\n        \"sys\": 120,\r\n        \"dias\": 80,\r\n        \"pulse\": 90,\r\n        \"timestamp\": 1652119779,\r\n        \"userID\": 3\r\n    }",
                "parameters": [
                    {
                        "in": "body",
                        "name": "vitals",
                        "description": "Vital parameters for a user.",
                        "schema": {
                            "$ref": "#/definitions/Vitals"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Post parameters not defined."
                    },
                    "200": {
                        "description": "Vitals successfully added."
                    }
                },
                "produces": [
                    "application/json"
                ]
            }
        },
        "/updateVitals": {
            "put": {
                "tags": [
                    "Vitals"
                ],
                "summary": "Update vital parameters for specific user.",
                "description": "Sample request:\r\n    \r\n    PUT /Gateway\r\n    {\r\n        \"sys\": 120,\r\n        \"dias\": 80,\r\n        \"pulse\": 90,\r\n        \"timestamp\": 1652119779,\r\n        \"userID\": 3\r\n    }",
                "parameters": [
                    {
                        "in": "body",
                        "name": "vitals",
                        "description": "Vital parameters for a user.",
                        "schema": {
                            "$ref": "#/definitions/Vitals"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Put parameters not defined."
                    },
                    "200": {
                        "description": "Vitals successfully updated."
                    }
                },
                "produces": [
                    "application/json"
                ]
            }
        },
        "/deleteVitals": {
            "delete": {
                "tags": [
                    "Vitals"
                ],
                "summary": "Delete vital parameters for specific user.",
                "parameters": [
                    {
                        "name": "userID",
                        "type": "integer",
                        "in": "query",
                        "description": "User ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vitals successfully deleted."
                    }
                }
            }
        }
    },
    "definitions": {
          "Vitals": {
            "type": "object",
            "properties": {
              "sys": {
                "type": "integer",
                "format": "int32",
                "default": 120
              },
              "dias": {
                "type": "integer",
                "format": "int32",
                "default": 80
              },
              "pulse": {
                "type": "integer",
                "format": "int32",
                "default": 70
              },
              "timestamp": {
                "type": "integer",
                "format": "int64",
                "default": 1234567890123
              },
              "userID": {
                "type": "integer",
                "format": "int32",
                "default": 1
              }
            
            }
        }
    }
} 